refactor(cascade-api): move Id type to separate module
