+++ cascade-api/src/Cascade/Api/Data/Prelude.hs
@@ -12,14 +12,6 @@ Portability : POSIX
 
 module Cascade.Api.Data.Prelude where
 
-import           Control.Lens.TH                ( makeWrapped )
-import           Data.Aeson                     ( FromJSON
-                                                , ToJSON
-                                                )
-import           Servant.API                    ( FromHttpApiData
-                                                , ToHttpApiData
-                                                )
-
 data family Readable (a :: Type)
 
 data family Creatable (a :: Type)
@@ -27,10 +19,3 @@ data family Creatable (a :: Type)
 data family Updatable (a :: Type)
 
 data family Deletable (a :: Type)
-
-newtype Id (entity :: Type) = Id
-  { unId :: UUID }
-  deriving stock Generic
-  deriving newtype (Show, Eq, Ord, FromHttpApiData, ToHttpApiData, FromJSON, ToJSON)
-
-makeWrapped ''Id
