+++ cascade-api/test/Cascade/Api/Hedgehog/Gen/Id.hs
@@ -2,9 +2,9 @@ module Cascade.Api.Hedgehog.Gen.Id
   ( id
   ) where
 
-import           Cascade.Api.Data.Prelude       ( Id(..) )
+import           Cascade.Api.Data.Id
 import qualified Cascade.Api.Hedgehog.Gen      as Gen
 import           Hedgehog
 
 id :: MonadGen m => m (Id entity)
-id = Gen.uuid |> fmap Id
+id = Id <$> Gen.uuid
